wot-servient/src/test/java/city/sane/wot/thing/TypeTest.java
/*
 * Copyright (c) 2021.
 *
 * This file is part of SANE Web of Things Servient.
 *
 * SANE Web of Things Servient is free software: you can redistribute it and/or
 * modify it under the terms of the GNU Lesser General Public License as
 * published by the Free Software Foundation, either version 3 of the License,
 * or (at your option) any later version.
 *
 * SANE Web of Things Servient is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU Lesser General Public License
 * along with SANE Web of Things Servient.  If not, see
 * <http://www.gnu.org/licenses/>.
 */
package city.sane.wot.thing;

import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.ObjectMapper;
import net.javacrumbs.jsonunit.core.Option;
import org.junit.jupiter.api.Test;

import java.io.IOException;

import static net.javacrumbs.jsonunit.assertj.JsonAssertions.assertThatJson;
import static org.junit.jupiter.api.Assertions.assertEquals;

class TypeTest {
    private final ObjectMapper jsonMapper = new ObjectMapper();

    @Test
    public void fromJson() throws IOException {
        // single value
        assertEquals(
                new Type("Thing"),
                jsonMapper.readValue("\"Thing\"", Type.class)
        );

        // array
        assertEquals(
                new Type("Thing").addType("saref:LightSwitch"),
                jsonMapper.readValue("[\"Thing\",\"saref:LightSwitch\"]", Type.class)
        );
    }

    @Test
    public void toJson() throws JsonProcessingException {
        // single value
        assertEquals(
                "\"Thing\"",
                jsonMapper.writeValueAsString(new Type("Thing"))
        );

        // multi type array
        assertThatJson(jsonMapper.writeValueAsString(new Type("Thing").addType("saref:LightSwitch")))
                .when(Option.IGNORING_ARRAY_ORDER)
                .isArray()
                .contains("Thing", "saref:LightSwitch");
    }
}